+++ src/test_old/app.py
@@ -13,13 +13,14 @@ from loguru import logger
 
 from graia.ariadne.entry import *
 from graia.ariadne.message.parser.base import RegexGroup
+from graia.ariadne.util import RichLogInstallOptions
 
 if __name__ == "__main__":
     url, account, verify_key, target, t_group = (
         open(os.path.join(__file__, "..", "test.temp"), "r").read().split(" ")
     )
     ALL_FLAG = True
-    Ariadne.config(inject_bypass_listener=True, install_log=True)
+    Ariadne.config(inject_bypass_listener=True, install_log=RichLogInstallOptions(rich_traceback=True))
 
     app = Ariadne(
         config(
@@ -194,11 +195,11 @@ if __name__ == "__main__":
     async def regex(app: Ariadne, chain: Annotated[MessageChain, RegexGroup("args")]):
         await app.send_friend_message(target, chain)
 
-    @bcc.receiver(ApplicationLaunched)
+    @bcc.receiver(ApplicationLaunch)
     async def m(app: Ariadne):
         await app.send_friend_message(target, MessageChain("Launched!"))
 
-    @bcc.receiver(ApplicationShutdowned)
+    @bcc.receiver(ApplicationShutdown)
     async def m(app: Ariadne):
         await app.send_friend_message(target, MessageChain("Shutdown!"))
 
@@ -206,7 +207,7 @@ if __name__ == "__main__":
     async def cmd_log(event: CommandExecutedEvent):
         devtools.debug(event)
 
-    @bcc.receiver(ApplicationLaunched)
+    @bcc.receiver(ApplicationLaunch)
     async def main():
         logger.debug(await app.get_version())
         logger.debug(await app.get_bot_profile())