+++ src/graia/ariadne/event/__init__.py
@@ -5,10 +5,9 @@ from pydantic import validator
 from ..dispatcher import BaseDispatcher
 from ..exception import InvalidEventTypeDefinition
 from ..model import AriadneBaseModel
-from ..util import AttrConvertMixin
 
 
-class MiraiEvent(Dispatchable, AriadneBaseModel, AttrConvertMixin):
+class MiraiEvent(Dispatchable, AriadneBaseModel):
     """Ariadne 的事件基类"""
 
     type: str