+++ src/graia/ariadne/event/mirai.py
@@ -1,14 +1,12 @@
 """Mirai 的各种事件"""
 from datetime import datetime
 from enum import Enum
-from typing import TYPE_CHECKING, Any, Dict, List, Optional
+from typing import Any, Dict, List, Optional
 
 from graia.broadcast.entities.dispatcher import BaseDispatcher
 from graia.broadcast.interfaces.dispatcher import DispatcherInterface
 from pydantic import Field
 from typing_extensions import Literal
-
-from graia.ariadne.util import deprecated
 
 from ..connection.util import CallMethod
 from ..message.chain import MessageChain
@@ -221,13 +219,6 @@ class BotMuteEvent(GroupEvent, BotEvent):
 
     operator: Member
     """执行禁言操作的管理员/群主"""
-
-    if not TYPE_CHECKING:
-
-        @property
-        @deprecated("0.8.0", "Use `duration` instead")
-        def duration_seconds(self) -> int:
-            return self.duration
 
     class Dispatcher(BaseDispatcher):
         @staticmethod
@@ -927,13 +918,6 @@ class MemberMuteEvent(GroupEvent):
     type = "MemberMuteEvent"
     duration: int = Field(..., alias="durationSeconds")
     """禁言时长, 单位为秒"""
-
-    if not TYPE_CHECKING:
-
-        @property
-        @deprecated("0.8.0", "Use `duration` instead")
-        def duration_seconds(self):
-            return self.duration
 
     member: Member
     """被禁言的成员"""